[Type checker] Keep a type checker alive during name binding.
Name lookup can cause import of some Clang declaration, which in turn
need a type checker. Make sure a type checker exists at this time.

This is a baby step toward having the type checker always available,
which is really the end goal. For now, we need this for better generic
environment (de-)serialization.