Update branches for new master-next scheme.
Combine the "next" and "next-upstream" checkout schemes since the new
approach for maintaining master-next is to building against the
"upstream-with-swift" Clang/LLVM branches. We do not plan to use the
"stable-next" branches anymore. (We can delete those branches after
this transition has had time to settle out.) I have also merged the
"upstream" scheme with "next-upstream" because building the combination
of branches previously in the scheme is not something we expect to work
now.