Fix a bug in FunctionSignatureOptimization.
When FSO generates a unique mangled name it should check that it does not conflict with any symbol in the current module or any public symbol from other modules.

FSO seems to have more issues with resilience, but those are not handled in this PR. They will be covered later. This PR is only for unblocking the inlining of generics.