function signature opts: Have an even more stricter check if a specialized function with that name already exists.
It’s not enough to just check for public functions in other modules.
In case of fragile functions, it can also be a non-public function in another module which potentially interferes with the specialization we want to create.

Anyway, this is just a workaround until we fix the mangling for function-signature opts. It should always produce unique names.