AST: Fix GenericSignatureBuilder bug with protocol typealiases
If a nested type of a generic parameter was a typealias, and the
generic parameter was not at depth 0, index 0, we would resolve
the type down to the wrong PotentialArchetype, causing bogus
diagnostics and crashes.

Make sure we apply a substitution to the typealias's underlying
type before resolving it, to map it to the correct PotentialArchetype.

This is a back-port of a similar fix that went into master.

Note that some parts of the test case don't typecheck in 3.1,
because of subsequent fixes that went into master, but the original
project in the radar now compiles.

Fixes <rdar://problem/30248571>.