[TypeCheck] Introduce a minimal validation that does enough for name lookup.
There was an underlying recursion between name lookup and validation
which is only exposed by where clauses on associated types, where
looking up an associated type required the protocol to be validated, but
validating a protocol requires any requirements (in where clauses) to be
validated. The latter may of course include references to associated
types, which is problematic. Fortunately, validateDecl is fundamentally
doing more work than necessary for name lookup, so we can break the
recursion by just separating out that minimal work.

At the moment, associated types and protocols are the source of the
recursion, so they're the only ones separated. They just need their
interface types (which is trivial for both) and accessibility (also
simple) for name lookup, along with any inherited protocols for a
protocol.