[silgen] Remove the +0 self hack.
The output from this change is the same except that:

1. Certain casts in SILGen do not create new cleanups and just propagate the
cleanup of the value that they are consuming. I have a later commit that is
going to fix this.

2. In general we should fully evaluate the result of a function before we
destroy any of its arguments. Previously with the +0 self hack, we were not
doing this. Now we do this properly.

rdar://29791263