[sil-cloner]  Fix a bug related to remapping of opened existentials during SIL function cloning
Also take the opportunity to simplify the code and tighten the assertions.
The commit does not contain a test-case, because it is kind of difficult to provide a reduced one.

The idea of the fix stems from @slavapestov. I just provided minor improvements and tested it on a very big test-case to see that it fixes the bug.